requirements.txt
torch
ffmpeg-python
torchaudio
opencv-python-headless
numpy
scikit-learn
transformers
